CRUD de Campeonato v2 funcionando

+++ src/main/java/com/agon/tcc/service/CampeonatoService.java
@@ -2,6 +2,7 @@ package com.agon.tcc.service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import javax.management.RuntimeErrorException;
 
@@ -9,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.agon.tcc.dto.CampeonatoDTO;
 import com.agon.tcc.model.Campeonato;
 import com.agon.tcc.repository.CampeonatoRepository;
 
@@ -18,38 +20,48 @@ public class CampeonatoService {
 	@Autowired
 	private CampeonatoRepository campeonatoRepository;
 
-	public List<Campeonato> findAll() {
-		return this.campeonatoRepository.findAll();
+	public List<CampeonatoDTO> findAll() {
+		return campeonatoRepository.findAll()
+				.stream()
+				.map(c -> new CampeonatoDTO(c.getId(), c.getNome(), c.getQuantidadeEquipes(), c.getDataInicio(), c.getDataFim(), 
+											c.getRegulamento(), c.getImagemCampeonato(), c.getFormato(), c.getModalidade()))
+				.collect(Collectors.toList());
 	}
 	
-	public Campeonato findById(Long id) {
-		Optional<Campeonato> campeonato = this.campeonatoRepository.findById(id);
-		return campeonato.orElseThrow(() -> new RuntimeException("Campeonato " + id + " não encontrado!"));
-	}
-	
-	@Transactional
-	public Campeonato create(Campeonato camp) {
-		camp.setId(null);
-		camp = campeonatoRepository.save(camp);
-		return camp;
-	}
-	
-	@Transactional
-	public Campeonato update(Campeonato camp) {
-		Campeonato newCamp = findById(camp.getId());
-		newCamp.setNome(camp.getNome());
-		newCamp.setDataInicio(camp.getDataInicio());
-		newCamp.setQuantidadeEquipes(camp.getQuantidadeEquipes());
-		return this.campeonatoRepository.save(newCamp);
-	}
-	
-	public void delete(Long id) {
-		findById(id);
-		try {
-			this.campeonatoRepository.deleteById(id);
-		} catch (Exception e) {
-			throw new RuntimeErrorException(null, "Não é possível excluir pois há entidades relacionadas!");
+	public CampeonatoDTO findById(Long id) {
+		Optional<Campeonato> campeonato = campeonatoRepository.findById(id);
+		if (campeonato.isPresent()) {
+			Campeonato c = campeonato.get();
+			return new CampeonatoDTO(c.getId(), c.getNome(), c.getQuantidadeEquipes(), c.getDataInicio(), c.getDataFim(), 
+									 c.getRegulamento(), c.getImagemCampeonato(), c.getFormato(), c.getModalidade());
+		} else {
+			return null;
 		}
 	}
 	
+	@Transactional
+	public void create(CampeonatoDTO campeonatoDTO) {
+		campeonatoRepository.save(new Campeonato(campeonatoDTO));
+	}
+	
+	@Transactional
+	public void update(CampeonatoDTO campeonatoDTO) {
+		Campeonato novoCamp = new Campeonato(findById(campeonatoDTO.id()));
+		novoCamp.setNome(campeonatoDTO.nome());
+		novoCamp.setDataInicio(campeonatoDTO.dataInicio());
+		novoCamp.setDataFim(campeonatoDTO.dataFim());
+		novoCamp.setQuantidadeEquipes(campeonatoDTO.quantidadeEquipes());
+		this.campeonatoRepository.save(novoCamp);
+	}
+	
+	public void delete(Long id) {
+		if (findById(id) != null) {
+			try {
+				this.campeonatoRepository.deleteById(id);
+			} catch (Exception e) {
+				throw new RuntimeErrorException(null, "Não é possível excluir pois há entidades relacionadas!");
+			}
+		}
+	}
+		
 }
